RUBY-214 - JRuby timestamps have millisecond precision and can cause client timestamp collisions. * Various moving/renaming of classes to clarify their use.
