+++ lib/cassandra/timestamp_generator/ticking_on_duplicate.rb
